Made content page more generic - Refactored routes to have one category view for routing - Refactored content page to load content from api to be generic - Cleaned up and fixed spelling in db.json
