+++ src/views/CategoryRouter.vue
